build(make): GNU make 4.3 is introduced to be the build tool.
The whole project is reorganized: source code and heading files are stored in directory `src` and `include` respectively.
All the generated files are stored in directory build; the executable file is located in `build\bin` and should be
executed with the rest of `build\bin` staying in the same directory.